Update automotive demo to match 1.0 HW mappings
This includs updating the addresses of GPIO ranges used, changing the
input GPIO mappings for the joystick to match the new GPIO input
mappings after pinmux, now using PMOD0 pin 1 instead of the Mikrobus
Interrupt as a GPIO digital input for the digital pedal demo (as that is
no longer mapped to be anything), and changing the legacy ethernet and
LCD driver code to now use SPI Chip selects instead of GPIO pins.

+++ examples/automotive/lib/analogue_pedal.c
@@ -112,7 +112,7 @@ static void analogue_task_one()
 bool analogue_task_two()
 {
 	// Control the volume bar via joystick input
-	uint8_t joystick = callbacks.joystick_read();
+	uint16_t joystick = callbacks.joystick_read();
 	if (joystick_in_direction(joystick, Up) && taskTwoMem->volume > 0)
 	{
 		taskTwoMem->volume -= 1;
